Move URLUtils to Client package.
